Overhaul of map/cache sinks to make use of better parallelism
